魔王血條寫法改版和魔王受傷
1.把魔王血條血法改成跟小怪依樣
2.魔王現在會受傷了

+++ ActionGameProject/Assets/ActionGame/UI/enemy/Health.cs
@@ -25,7 +25,7 @@ public class Health : MonoBehaviour
     }
 
     //血量增減顯示
-    public void ModifyHealth(int amout)
+    public void ModifyHealth(float amout)
     {
         currentHealth -= amout;
